Refactor CLI::Application
- Merge CLI::Input into CLI::Application
- Move all arguments for commands to BaseCommand's initializer
- Remove OptionInterpreter by moving its methods to ReportCommand
