+++ lib/reek/cli/application.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 require_relative 'options'
-require_relative 'option_interpreter'
 require_relative '../configuration/app_configuration'
+require_relative '../source/source_locator'
 require_relative 'command/report_command'
 require_relative 'command/todo_list_command'
 
@@ -13,18 +13,19 @@ module Reek
     # command line.
     #
     class Application
-      include Input
       attr_reader :configuration
 
       def initialize(argv)
         @options = configure_options(argv)
         @status = options.success_exit_code
         @configuration = configure_app_configuration(options.config_file)
-        @command = command_class.new(OptionInterpreter.new(options), sources: sources)
+        @command = command_class.new(options: options,
+                                     sources: sources,
+                                     configuration: configuration)
       end
 
       def execute
-        command.execute self
+        command.execute
       end
 
       private
@@ -50,8 +51,43 @@ module Reek
         options.generate_todo_list ? Command::TodoListCommand : Command::ReportCommand
       end
 
+      def sources
+        if no_source_files_given?
+          if input_was_piped?
+            source_from_pipe
+          else
+            working_directory_as_source
+          end
+        else
+          sources_from_argv
+        end
+      end
+
       def argv
         options.argv
+      end
+
+      # :reek:UtilityFunction
+      def input_was_piped?
+        !$stdin.tty?
+      end
+
+      def no_source_files_given?
+        # At this point we have deleted all options from argv. The only remaining entries
+        # are paths to the source files. If argv is empty, this means that no files were given.
+        argv.empty?
+      end
+
+      def working_directory_as_source
+        Source::SourceLocator.new(['.'], configuration: configuration).sources
+      end
+
+      def sources_from_argv
+        Source::SourceLocator.new(argv).sources
+      end
+
+      def source_from_pipe
+        [$stdin]
       end
     end
   end
