+++ lib/reek/cli/command/report_command.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 require_relative 'base_command'
 require_relative '../../examiner'
+require_relative '../../report'
 
 module Reek
   module CLI
@@ -10,19 +11,19 @@ module Reek
       # text report format.
       #
       class ReportCommand < BaseCommand
-        def execute(app)
-          populate_reporter_with_smells app
+        def execute
+          populate_reporter_with_smells
           reporter.show
           result_code
         end
 
         private
 
-        def populate_reporter_with_smells(app)
+        def populate_reporter_with_smells
           sources.each do |source|
             reporter.add_examiner Examiner.new(source,
                                                filter_by_smells: smell_names,
-                                               configuration: app.configuration)
+                                               configuration: configuration)
           end
         end
 
@@ -31,7 +32,36 @@ module Reek
         end
 
         def reporter
-          @reporter ||= options.reporter
+          @reporter ||=
+            report_class.new(
+              warning_formatter: warning_formatter,
+              report_formatter: Report::Formatter,
+              sort_by_issue_count: sort_by_issue_count,
+              heading_formatter: heading_formatter)
+        end
+
+        def report_class
+          Report.report_class(options.report_format)
+        end
+
+        def warning_formatter
+          warning_formatter_class.new(location_formatter: location_formatter)
+        end
+
+        def warning_formatter_class
+          Report.warning_formatter_class(options.show_links ? :wiki_links : :simple)
+        end
+
+        def location_formatter
+          Report.location_formatter(options.location_format)
+        end
+
+        def heading_formatter
+          Report.heading_formatter(options.show_empty ? :verbose : :quiet)
+        end
+
+        def sort_by_issue_count
+          options.sorting == :smelliness
         end
       end
     end
