+++ spec/factories/factories.rb
@@ -1,7 +1,6 @@
 require_relative '../../lib/reek/smells'
 require_relative '../../lib/reek/smells/smell_detector'
 require_relative '../../lib/reek/smells/smell_warning'
-require_relative '../../lib/reek/cli/option_interpreter'
 require_relative '../../lib/reek/cli/options'
 
 FactoryGirl.define do
@@ -51,16 +50,6 @@ FactoryGirl.define do
           lines: lines,
           message: message,
           parameters: parameters)
-    end
-  end
-
-  factory :options_interpreter_with_empty_sources, class: Reek::CLI::OptionInterpreter do
-    transient do
-      options { Reek::CLI::Options.new [] }
-    end
-
-    initialize_with do
-      Reek::CLI::OptionInterpreter.new(options)
     end
   end
 end
